refactor: move attn_ffn_piecewise to atom/utils, trim comments
Cudagraph plumbing, not a model op -- it belongs beside the runner it wraps.
Also fixes a stale reference to run_attn_ffn_piecewise, now a method.
